# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation Test Indosat",
        "contact": {
            "email": "jauharimalikupil@gmail.com"
        },
        "license": {
            "name": "GPL2",
            "url": "https://jauharimalik.github.io"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoice": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get invoice List",
                "description": "Get invoice List as Array",
                "operationId": "index invoice",
                "responses": {
                    "200": {
                        "description": "Get invoice List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/invoice/show/{id}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Show Invoice Details",
                "description": "Show Invoice Details",
                "operationId": "Show Invoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Invoice Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/invoice/create": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Create New Invoice",
                "description": "Create New Invoice",
                "operationId": "Create Invoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 104
                                    },
                                    "duedate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "27-07-2023"
                                    },
                                    "details": {
                                        "type": "object",
                                        "example": [
                                            {
                                                "id": 1,
                                                "harga": 15000,
                                                "qty": 3
                                            },
                                            {
                                                "id": 2,
                                                "harga": 35000,
                                                "qty": 5
                                            },
                                            {
                                                "id": 3,
                                                "harga": 38000,
                                                "qty": 9
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Invoice"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/invoice/edit": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Edit Invoice",
                "description": "Edit Invoice",
                "operationId": "Edit Invoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomorbukti": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 104
                                    },
                                    "duedate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "27-07-2023"
                                    },
                                    "details": {
                                        "type": "object",
                                        "example": [
                                            {
                                                "id": 1,
                                                "harga": 15000,
                                                "qty": 3
                                            },
                                            {
                                                "id": 2,
                                                "harga": 35000,
                                                "qty": 5
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Data Berhasil"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/invoice/{id}": {
            "delete": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Delete Invoice",
                "description": "Delete Invoice",
                "operationId": "destroy invoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Invoice"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Product Details",
                "description": "Show Product Details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Product Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Update Product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Product 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10120
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "description": "Delete Product",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Product"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}